# kvm: add a few sanity checks on vmexit-s