--- a/pkg/ring0/defs_arm64.go
+++ b/pkg/ring0/defs_arm64.go
@@ -98,8 +98,10 @@ func (c *CPU) ClearErrorCode() {
 	c.errorType = 1 // User mode.
 }
 
+// FaultAddr returns the last fault address.
+//
 //go:nosplit
-func (c *CPU) GetFaultAddr() (value uintptr) {
+func (c *CPU) FaultAddr() (value uintptr) {
 	return c.faultAddr
 }