Here are not visible perfomance effects:

Before:
```
goos: linux
goarch: amd64
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
BenchmarkApplicationSyscall
I0401 16:23:28.351151  2594711 physical_map.go:145] region: virtual [3f1eaf85f000,7f202f85f000)
I0401 16:23:28.351260  2594711 physical_map.go:197] physicalRegion: virtual [1000,3f1eaf85f000) => physical [100001000,3f1faf85f000)
I0401 16:23:28.351287  2594711 physical_map.go:197] physicalRegion: virtual [7f202f85f000,7ffffffff000) => physical [3f1faf85f000,3fff7ffff000)
BenchmarkApplicationSyscall-6           	 2574579	       408.2 ns/op
BenchmarkKernelSyscall
BenchmarkKernelSyscall-6                	 2268025	       500.0 ns/op
BenchmarkKernelVDSO
BenchmarkKernelVDSO-6                   	26895987	        41.30 ns/op
BenchmarkWorldSwitchToUserRoundtrip
BenchmarkWorldSwitchToUserRoundtrip-6   	  188020	      5655 ns/op
```

After:
```
goos: linux
goarch: amd64
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
BenchmarkApplicationSyscall
I0401 16:23:56.143275  2595618 physical_map.go:145] region: virtual [3f5ec46d0000,7f60446d0000)
I0401 16:23:56.143387  2595618 physical_map.go:197] physicalRegion: virtual [1000,3f5ec46d0000) => physical [100001000,3f5fc46d0000)
I0401 16:23:56.143403  2595618 physical_map.go:197] physicalRegion: virtual [7f60446d0000,7ffffffff000) => physical [3f5fc46d0000,3fff7ffff000)
BenchmarkApplicationSyscall-6           	 2798324	       409.8 ns/op
BenchmarkKernelSyscall
BenchmarkKernelSyscall-6                	 2314084	       490.8 ns/op
BenchmarkKernelVDSO
BenchmarkKernelVDSO-6                   	26801140	        40.11 ns/op
BenchmarkWorldSwitchToUserRoundtrip
BenchmarkWorldSwitchToUserRoundtrip-6   	  202519	      5739 ns/op
```